+++ aeron-client/src/test/cpp/ClientConductorFixture.h
@@ -43,7 +43,7 @@ using namespace std::placeholders;
 
 static const long DRIVER_TIMEOUT_MS = 10 * 1000;
 static const long RESOURCE_LINGER_TIMEOUT_MS = 5 * 1000;
-static const long INTER_SERVICE_TIMEOUT_NS = 5 * 1000 * 1000 * 1000L;
+static const long long INTER_SERVICE_TIMEOUT_NS = 5 * 1000 * 1000 * 1000LL;
 static const long INTER_SERVICE_TIMEOUT_MS = INTER_SERVICE_TIMEOUT_NS / 1000000L;
 static const long PUBLICATION_CONNECTION_TIMEOUT_MS = 5 * 1000L;
